+++ springdoc-openapi-tests/springdoc-openapi-kotlin-data-rest-tests/src/test/kotlin/test/org/springdoc/api/v31/app1/PersonRepository.kt
@@ -1,0 +1,7 @@
+package test.org.springdoc.api.v31.app1
+
+import org.springframework.data.repository.PagingAndSortingRepository
+import org.springframework.data.rest.core.annotation.RepositoryRestResource
+
+@RepositoryRestResource(collectionResourceRel = "people", path = "people")
+interface PersonRepository : PagingAndSortingRepository<Person, Long>